Part5_AI_Agent_프로젝트_3개/worktree-cultural-reservation/Part2_Agent_개념과_아키텍처/Chapter2_LLM_API_호출_이해하기/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
